auth-service/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-easyjwt==0.5.2
djangorestframework==3.15.0
idna==3.6
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
